[ALPS05652867] mml: provide new adaptor and new config flow
- For disp addon mode, define DLE (MML as a Direct-LinkeE) adaptor
  to support its passive control flow.
- In a passive connection, do not make command nor flush cmdq_pkt
  from MML driver.
- To reduce redundant flow, provide frame mode from tile calculation.
- Add addon path into platform topology.

MTK-Commit-Id: d4fcada4ac24585feaabb1d07529dda2dbfb8f4d

CR-Id: ALPS05652867
Feature: [Module]MML driver

+++ drivers/gpu/drm/mediatek/mml/mtk-mml-tile.h
@@ -84,6 +84,6 @@ s32 calc_tile(struct mml_task *task, u32 pipe_idx);
 
 void destroy_tile_output(struct mml_tile_output *output);
 
-void dump_tile_output(struct mml_task *task, u8 pipe_idx);
+void dump_tile_output(struct mml_task *task, u32 pipe_idx);
 
 #endif	/* __MTK_MML_TILE_H__ */
